+++ src/edit/EditToolbar.js
@@ -3,10 +3,21 @@
 });*/
 
 L.EditToolbar = L.Toolbar.extend({
+	statics: {
+		TYPE: 'edit'
+	},
+
 	options: {
 		edit: {
 			selectedPathOptions: {
 				dashArray: '10, 10',
+
+				fill: true,
+				fillColor: '#fe57a1',
+				fillOpacity: 0.1,
+
+				// Whether to user the existing layers color
+				maintainColor: false
 			}
 		},
 		remove: {},
@@ -19,7 +30,7 @@ L.EditToolbar = L.Toolbar.extend({
 			if (typeof options.edit.selectedPathOptions === 'undefined') {
 				options.edit.selectedPathOptions = this.options.edit.selectedPathOptions;
 			}
-			options.edit = L.extend({}, this.options.edit, options.edit);
+			options.edit.selectedPathOptions = L.extend({}, this.options.edit.selectedPathOptions, options.edit.selectedPathOptions);
 		}
 
 		if (options.remove) {