+++ CHANGELOG.md
@@ -10,6 +10,7 @@ An in-progress version being developed on the master branch.
 ### Improvements
 
  * Add support for toolbar touch styles. (by [@alanshaw](https://github.com/alanshaw)). [#269](https://github.com/Leaflet/Leaflet.draw/pull/269 )
+ * Add support for maintaining a layers color when entering edit mode.
 
 ## 0.2.3 (January 14, 2014)